Remove warnings generated by tests.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@33686 72102866-910b-0410-8b05-ffd578937521

+++ Tests/base/GenericTests/generic.m
@@ -9,7 +9,7 @@
 {
   return ([self isEqual:anObject] && [anObject isEqual:self]);
 }
-- (int) length
+- (NSUInteger) length
 {
   return 0;
 }
